Import parsing and composing from domain. Restructure.
