+++ Cargo.toml
@@ -1,7 +1,8 @@
 [package]
 name = "octets"
 version = "0.1.0"
-edition = "2018"
+edition = "2021"
+rust_version = "0.56"
 authors = ["Martin Hoffmann <martin@nlnetlabs.nl>"]
 description = "Abstractions for types representing octet sequences."
 documentation = "https://docs.rs/octets"
